etl/requirements.txt
kafka-python
redis 